refactor(core): extraer helpers de marketing-sections, industry-gallery y ContentPanel
- marketing-sections-manager.tsx: extraer SECTION_LABELS, SECTION_ORDER,
  SECTION_FIELDS, tipos y estilos de input a marketing-sections-helpers.ts
  (789→611, -178 líneas)
- industry-gallery-manager.tsx: extraer extractErrorMessage, CardFormState
  y EMPTY_CARD_FORM a industry-gallery-helpers.ts (755→717, -38 líneas)
- ContentPanel.tsx: extraer SectionContent y ContentPanelProps interfaces
  a content-panel-types.ts (657→619, -38 líneas)

Part of #244
